refactor: Refactor `GpuFence` into a proper RAII wrapper (inherits `GpuResource`, deferred disposal)

+++ VanillaGraphicsExpanded.Tests/GPU/GpuEboIntegrationTests.cs
@@ -603,7 +603,7 @@ public class GpuEboIntegrationTests
         DrawElementsType originalIndexType = ebo.IndexType;
 
         // Act
-        int detachedId = ebo.Detach();
+        int detachedId = (int)ebo.Detach();
 
         // Assert
         Assert.Equal(originalBufferId, detachedId);
